particul/requirements.txt
torch >= 1.09
torchvision >= 0.8.2
numpy >= 1.19.5
pillow == 9.0.0
progress == 1.5
scikit-image >= 0.18.1
scipy~=1.8.1
setuptools~=60.2.0
matplotlib~=3.5.2
pandas~=1.4.3
scikit-learn~=1.1.1